Split ActionForm and BaseForm storybook stories into focused files
Break up the large ActionForm.stories.tsx and BaseForm.stories.tsx into
per-topic story files (Fields, Title, States/Submission, Validation,
Dialog, TitleAndSections) and extract shared setup into
actionFormStoryHelpers and baseFormStoryHelpers.
