+++ packages/react-components-storybook/src/stories/ActionForm/ActionForm.Fields.stories.tsx
@@ -1,0 +1,197 @@
+/*
+ * Copyright 2026 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { Meta, StoryObj } from "@storybook/react-vite";
+
+import {
+  actionFormDefaultValueFields,
+  actionFormMeta,
+  actionFormOverrideFields,
+  type UpdateEmployeeActionFormStoryProps,
+  UnsupportedFieldsActionFormStory,
+} from "./actionFormStoryHelpers.js";
+
+const meta: Meta<UpdateEmployeeActionFormStoryProps> = {
+  ...actionFormMeta,
+  title: "Components/ActionForm/Features/Fields",
+};
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const WithDefaultValues: Story = {
+  args: {
+    formFieldDefinitions: actionFormDefaultValueFields,
+  },
+  parameters: {
+    docs: {
+      source: {
+        code: `const fields = [
+  {
+    fieldKey: "fullName",
+    label: "Full name",
+    fieldComponent: "TEXT_INPUT",
+    defaultValue: "Grace Hopper",
+    fieldComponentProps: {},
+  },
+  {
+    fieldKey: "yearsExperience",
+    label: "Years of experience",
+    fieldComponent: "NUMBER_INPUT",
+    defaultValue: 42,
+    fieldComponentProps: { min: 0 },
+  },
+  {
+    fieldKey: "isRemote",
+    label: "Remote employee",
+    fieldComponent: "RADIO_BUTTONS",
+    defaultValue: true,
+    fieldComponentProps: {
+      options: [
+        { label: "True", value: true },
+        { label: "False", value: false },
+      ],
+    },
+  },
+  {
+    fieldKey: "isFullTime",
+    label: "Employment type",
+    fieldComponent: "DROPDOWN",
+    defaultValue: true,
+    fieldComponentProps: {
+      items: [true, false],
+      itemToStringLabel: (value) =>
+        value === true ? "Full-time" : "Contractor",
+      placeholder: "Select employment type",
+    },
+  },
+] satisfies Array<FormFieldDefinition<typeof updateEmployeeStoryAction>>;
+
+<ActionForm
+  actionDefinition={updateEmployeeStoryAction}
+  formFieldDefinitions={fields}
+/>`,
+      },
+    },
+  },
+};
+
+export const WithFieldOverrides: Story = {
+  args: {
+    formFieldDefinitions: actionFormOverrideFields,
+  },
+  parameters: {
+    docs: {
+      source: {
+        code: `const fields = [
+  {
+    fieldKey: "fullName",
+    label: "Employee legal name",
+    fieldComponent: "TEXT_INPUT",
+    helperText: "Use the name that should appear in HR records.",
+    fieldComponentProps: {
+      placeholder: "Ada Lovelace",
+      minLength: 2,
+    },
+  },
+  {
+    fieldKey: "yearsExperience",
+    label: "Relevant experience",
+    fieldComponent: "NUMBER_INPUT",
+    helperText: "Whole years only.",
+    fieldComponentProps: {
+      min: 0,
+      max: 80,
+    },
+  },
+  {
+    fieldKey: "isRemote",
+    label: "Remote employee",
+    fieldComponent: "SWITCH",
+    helperText: "Turn on when the employee is not assigned to an office.",
+    fieldComponentProps: {},
+  },
+  {
+    fieldKey: "isFullTime",
+    label: "Employment type",
+    fieldComponent: "DROPDOWN",
+    helperText: "Choose whether this employee is full-time or a contractor.",
+    fieldComponentProps: {
+      items: [true, false],
+      itemToStringLabel: (value) =>
+        value === true ? "Full-time" : "Contractor",
+      placeholder: "Select employment type",
+    },
+  },
+] satisfies Array<FormFieldDefinition<typeof updateEmployeeStoryAction>>;
+
+<ActionForm
+  actionDefinition={updateEmployeeStoryAction}
+  formFieldDefinitions={fields}
+/>`,
+      },
+    },
+  },
+};
+
+export const WithUnsupportedFields: Story = {
+  render: () => <UnsupportedFieldsActionFormStory />,
+  parameters: {
+    docs: {
+      source: {
+        code: `// This story uses an action with parameter types that ActionForm does
+// not currently generate default field components for.
+//
+// {
+//   apiName: "unsupportedFieldsStoryAction",
+//   displayName: "Review unsupported fields",
+//   parameters: {
+//     structPayload: {
+//       displayName: "Struct payload",
+//       dataType: {
+//         type: "struct",
+//         fields: [
+//           {
+//             name: "externalId",
+//             fieldType: { type: "string" },
+//             required: true,
+//           },
+//         ],
+//       },
+//       required: true,
+//     },
+//     geoshape: {
+//       displayName: "Geoshape",
+//       dataType: { type: "geoshape" },
+//     },
+//     classification: {
+//       displayName: "Classification",
+//       dataType: { type: "marking" },
+//     },
+//     objectKind: {
+//       displayName: "Object type",
+//       dataType: { type: "objectType" },
+//     },
+//   },
+// }
+<ActionForm
+  actionDefinition={unsupportedFieldsStoryAction.actionDefinition}
+  showFormTitle={true}
+/>`,
+      },
+    },
+  },
+};
